Combine main scouting flow logic in one file and display message if there is no current competition on scouting flow page
