+++ app/src/FormHelper.js
@@ -2,8 +2,8 @@ import AsyncStorage from '@react-native-async-storage/async-storage';
 
 class FormHelper {
   static DEBUG = false;
-  static LATEST_FORM = '2023form';
-  static COMPETITION = '2023competition';
+  static LATEST_FORM = 'current-form';
+  static ASYNCSTORAGE_COMPETITION_KEY = 'current-competition';
   static SCOUTING_STYLE = 'scoutingStyle';
   static THEME = 'themePreference';
